extract useResources as a seperate file
