+++ hooks-simple/src/components/ResourceList.jsx
@@ -1,30 +1,6 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 
-import { resourceAPI } from "../apis/resource";
-
-const useResources = (resource) => {
-  const [resources, setResources] = useState([]);
-
-  const fetchResource = async (resource) => {
-    const response = await resourceAPI.get(resource);
-    setResources(response.data);
-  };
-
-  useEffect(() => {
-    fetchResource(resource);
-
-    /*
-     * This also works to bypass the limitation that useEffect callback cannot
-     * be asynchronous.
-     */
-    // (async (resource) => {
-    //   const response = await resourceAPI.get(resource);
-    //   setResources(response.data);
-    // })(resource);
-  }, [resource]);
-
-  return resources;
-};
+import { useResources } from "./useResources";
 
 const ResourceList = ({ resource }) => {
   const resources = useResources(resource);
